refactor: Make BuilderCanvas lighter and improve readability
- Move all canvas events to useCanvasEvents
- DropZone funtionality to useCanvasrDropZone
- All utilities to useCanvasrUtils
